storage: move SDKConnect/RawConnect to a package
