+++ storage/oasis/connections/raw_grpc.go
@@ -1,4 +1,4 @@
-package history
+package connections
 
 import (
 	"crypto/tls"
