+++ storage/oasis/connections/sdk.go
@@ -1,4 +1,4 @@
-package history
+package connections
 
 import (
 	"context"
